bugfixes addBook, tests
