+++ src/templates/base.html
@@ -26,7 +26,7 @@
 <body>
 <nav class="navbar">
     <div class="navbar__container">
-        <a href="#home" id="navbar__logo">HIDE AND READ
+        <a href="{{ url_for('mainpage') }}" id="navbar__logo">
         <img src='https://i.postimg.cc/FHCGgsMz/Logo3.png' width="50" height="50">
         </a>
         <div class="navbar__toggle" id="mobile-menu">
