+++ src/templates/content/map.html
@@ -134,7 +134,7 @@
         booknr = getBooks();
         console.log(booknr);
         // empty data object
-        if (typeof booknr == null) {
+        if (booknr == null) {
             alert('Wrong book number');
         }
         else {
@@ -150,10 +150,13 @@
                 data: JSON.stringify(data),
                 dataType: "json",
                 success: function(response) {
-                    console.log(response);
+                    console.log("success");
+                    alert(response);
+                    location.reload();
                 },
                 error: function(err) {
-                    console.log(err);
+                    console.log("error");
+                    alert(err);
                 }
             });
         }
@@ -170,7 +173,7 @@
         var book_ctr;
         {% if ubooks %}
             {% for id, book in ubooks.items() %}
-                prmpt_str = prmpt_str.concat("{{ book['id'] }} - {{ book['title'] }} - {{ book['author'] }} - {{ book['isbn'] }}");
+                prmpt_str = prmpt_str.concat("{{ id }} - {{ book['title'] }} - {{ book['author'] }} - {{ book['isbn'] }}");
                 {% if not loop.last %}
                     prmpt_str = prmpt_str.concat("\n");
                 {% endif %}
@@ -179,8 +182,11 @@
         {% endif %}
         console.log(prmpt_str);
         booknr = prompt(prmpt_str);
-        return booknr;
-        // TODO: fix for book number -> hier kann man alles rein hämmern :D
+        {% if booknr not in ubooks.keys() %}
+            return null;
+        {% else %}
+            return booknr;
+        {% endif %}
     }
